+++ notiolink.config.js
@@ -5,9 +5,19 @@ module.exports = {
   /** Description for SEO */
   seoDescription:
     'Self-hostable branded link shortener built with Next.js & Notion API',
+
   /** Twitter account for SEO */
   twitter: '',
 
   /** Without additional '/' on the end, e.g. https://theodorusclarence.com */
   deployUrl: 'https://jessica.vercel.app',
+
+  /**
+   * Color theme for the app
+   * @type {'light' | 'dark' | 'milky' | 'street' | 'monokai'}
+   */
+  theme: 'dark',
+
+  /** REQUIRED CONFIG: Set this to 'false'  */
+  demoMode: 'true',
 };