+++ src/components/forms/Input.tsx
@@ -1,7 +1,8 @@
-import clsx from 'clsx';
 import * as React from 'react';
 import { RegisterOptions, useFormContext } from 'react-hook-form';
 import { HiExclamationCircle } from 'react-icons/hi';
+
+import clsxm from '@/lib/clsxm';
 
 export type InputProps = {
   /** Input label */
@@ -46,10 +47,13 @@ export default function Input({
 
   return (
     <div className='w-full'>
-      <label htmlFor={id} className='block text-sm font-normal text-gray-200'>
+      <label
+        htmlFor={id}
+        className='block dark:text-gray-200 font-normal text-gray-700 text-sm'
+      >
         {label}
       </label>
-      <div className='relative mt-1'>
+      <div className='mt-1 relative'>
         <input
           {...register(id, validation)}
           {...rest}
@@ -57,28 +61,39 @@ export default function Input({
           name={id}
           id={id}
           readOnly={readOnly}
-          className={clsx(
-            readOnly
-              ? 'bg-gray-700 focus:ring-0 cursor-not-allowed border-gray-600 focus:border-gray-600'
-              : errors[id]
-              ? 'focus:ring-red-400 border-red-400 focus:border-red-400'
-              : 'focus:ring-primary-500 border-gray-600 focus:border-primary-500',
-            'bg-dark block w-full text-white rounded-md shadow-sm'
+          className={clsxm(
+            'bg-transparent',
+            'focus:border-primary-500 focus:ring-primary-500',
+            readOnly &&
+              'cursor-not-allowed bg-gray-700 focus:border-gray-600 focus:ring-0',
+            errors[id]
+              ? [
+                  'border-red-600 focus:border-red-600 focus:ring-red-600',
+                  'dark:border-red-400 dark:focus:border-red-400 dark:focus:ring-red-400',
+                ]
+              : ['border-gray-400', 'dark:border-gray-600'],
+            'block w-full rounded-md shadow-sm dark:text-white'
           )}
           placeholder={placeholder}
           aria-describedby={id}
         />
 
         {!hideError && errors[id] && (
-          <div className='flex absolute inset-y-0 right-0 items-center pr-3 pointer-events-none'>
-            <HiExclamationCircle className='text-xl text-red-400' />
+          <div className='absolute flex inset-y-0 items-center pointer-events-none pr-3 right-0'>
+            <HiExclamationCircle className='dark:text-red-400 text-red-600 text-xl' />
           </div>
         )}
       </div>
       <div className='mt-1'>
-        {helperText && <p className='text-xs text-gray-300'>{helperText}</p>}
+        {helperText && (
+          <p className='dark:text-gray-300 text-gray-600 text-xs'>
+            {helperText}
+          </p>
+        )}
         {!hideError && errors[id] && (
-          <span className='text-sm text-red-400'>{errors[id].message}</span>
+          <span className='dark:text-red-400 text-red-600 text-sm'>
+            {errors[id].message}
+          </span>
         )}
       </div>
     </div>