+++ src/components/layout/Header.tsx
@@ -9,13 +9,13 @@ const links = [
 
 export default function Header() {
   return (
-    <header className='sticky top-0 z-50 bg-white'>
-      <div className='layout flex justify-between items-center h-14'>
+    <header className='bg-white sticky top-0 z-50'>
+      <div className='flex h-14 items-center justify-between layout'>
         <UnstyledLink href='/' className='font-bold hover:text-gray-600'>
           Home
         </UnstyledLink>
         <nav>
-          <ul className='flex justify-between items-center space-x-4'>
+          <ul className='flex items-center justify-between space-x-4'>
             {links.map(({ href, label }) => (
               <li key={`${href}${label}`}>
                 <UnstyledLink href={href} className='hover:text-gray-600'>